+++ cmd/server/internal/handlers/accounts_test.go
@@ -1,17 +1,18 @@
 package handlers_test
 
 import (
+	"context"
+	"net/http"
+	"testing"
+
 	accountsv1 "buf.build/gen/go/xskydev/go-money-pb/protocolbuffers/go/gomoneypb/accounts/v1"
 	"connectrpc.com/connect"
-	"context"
 	"github.com/ft-t/go-money/cmd/server/internal/handlers"
 	"github.com/ft-t/go-money/cmd/server/internal/middlewares"
 	"github.com/ft-t/go-money/pkg/auth"
 	"github.com/ft-t/go-money/pkg/boilerplate"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"net/http"
-	"testing"
 )
 
 func TestReorder(t *testing.T) {
