+++ pkg/transactions/validation/validation_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ft-t/go-money/pkg/configuration"
 	"github.com/ft-t/go-money/pkg/database"
 	"github.com/ft-t/go-money/pkg/testingutils"
-	"github.com/ft-t/go-money/pkg/transactions"
+	"github.com/ft-t/go-money/pkg/transactions/applicable_accounts"
 	"github.com/ft-t/go-money/pkg/transactions/validation"
 	"github.com/golang/mock/gomock"
 	"github.com/shopspring/decimal"
@@ -69,7 +69,7 @@ func TestValidateWithdrawal(t *testing.T) {
 		})
 
 		applicableSvc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_EXPENSE: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[0],
@@ -172,7 +172,7 @@ func TestValidateWithdrawal(t *testing.T) {
 		})
 
 		appAcc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_EXPENSE: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[1],
@@ -344,7 +344,7 @@ func TestValidateDeposit(t *testing.T) {
 		})
 
 		accSvc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_INCOME: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[0],
@@ -458,7 +458,7 @@ func TestValidateReconciliation(t *testing.T) {
 		})
 
 		accSvc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_ADJUSTMENT: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[0],
@@ -608,7 +608,7 @@ func TestValidateTransferBetweenAccounts(t *testing.T) {
 		})
 
 		accSvc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_TRANSFER_BETWEEN_ACCOUNTS: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[0],
@@ -643,7 +643,7 @@ func TestValidateTransferBetweenAccounts(t *testing.T) {
 		})
 
 		accSvc.EXPECT().GetApplicableAccounts(gomock.Any(), gomock.Any()).
-			Return(map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+			Return(map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 				gomoneypbv1.TransactionType_TRANSACTION_TYPE_TRANSFER_BETWEEN_ACCOUNTS: {
 					SourceAccounts: map[int32]*database.Account{
 						acc[0].ID: acc[0],
@@ -692,7 +692,7 @@ func TestValidateTxAccounts(t *testing.T) {
 			ApplicableAccountSvc: accSvc,
 		})
 
-		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 			gomoneypbv1.TransactionType_TRANSACTION_TYPE_INCOME: {},
 		}, &database.Transaction{
 			TransactionType: gomoneypbv1.TransactionType_TRANSACTION_TYPE_EXPENSE,
@@ -709,7 +709,7 @@ func TestValidateTxAccounts(t *testing.T) {
 			ApplicableAccountSvc: accSvc,
 		})
 
-		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 			gomoneypbv1.TransactionType_TRANSACTION_TYPE_INCOME: {
 				SourceAccounts: map[int32]*database.Account{},
 			},
@@ -729,7 +729,7 @@ func TestValidateTxAccounts(t *testing.T) {
 			ApplicableAccountSvc: accSvc,
 		})
 
-		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 			gomoneypbv1.TransactionType_TRANSACTION_TYPE_INCOME: {
 				SourceAccounts: map[int32]*database.Account{
 					1: {},
@@ -753,7 +753,7 @@ func TestValidateTxAccounts(t *testing.T) {
 			ApplicableAccountSvc: accSvc,
 		})
 
-		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*transactions.PossibleAccount{
+		err := srv.ValidateTransactionAccounts(context.TODO(), map[gomoneypbv1.TransactionType]*applicable_accounts.PossibleAccount{
 			gomoneypbv1.TransactionType_TRANSACTION_TYPE_INCOME: {
 				SourceAccounts: map[int32]*database.Account{
 					1: {ID: 1, Currency: "USD"},
